重构创建文章

+++ app/models/article/article.go
@@ -6,12 +6,12 @@ import (
 )
 
 type Article struct {
-	ID    int
+	ID    int64
 	Title string
 	Body  string
 }
 
 // Link 方法用来生成文章链接
 func (a Article) Link() string  {
-	return route.Name2URL("articles.show", "id", strconv.FormatInt(int64(a.ID), 10))
+	return route.Name2URL("articles.show", "id", strconv.FormatInt(a.ID, 10))
 }
